+++ app/cogs/chatgpt.py
@@ -1,4 +1,3 @@
-import requests
 import json
 import aiohttp
