+++ app/routers/guild/line/line_set.py
@@ -1,20 +1,15 @@
-from fastapi import APIRouter,Header
+from fastapi import APIRouter
 from fastapi.responses import RedirectResponse,JSONResponse
 from starlette.requests import Request
 from fastapi.templating import Jinja2Templates
 
 from typing import List,Dict,Any
 
-from base.aio_req import (
-    aio_get_request,
-    return_permission,
-    discord_oauth_check,
-    get_profile,
-    sort_channels,
-    sort_discord_channel,
-    decrypt_password
-)
-
+from pkg.aio_req import aio_get_request
+from pkg.oauth_check import discord_oauth_check,discord_get_profile
+from pkg.permission import return_permission
+from pkg.sort_channel import sort_channels,sort_discord_channel
+from pkg.crypt import decrypt_password
 from model_types.discord_type.guild_permission import Permission
 from model_types.discord_type.discord_user_session import DiscordOAuthData
 from model_types.discord_type.discord_type import DiscordUser,Threads
@@ -244,7 +239,7 @@ class LineSetView(commands.Cog):
                 # アクセストークンの復号化
                 access_token = session.discord_oauth_data.access_token
                 # Discordのユーザ情報を取得
-                discord_user = await get_profile(access_token=access_token)
+                discord_user = await discord_get_profile(access_token=access_token)
 
                 # トークンが無効
                 if discord_user == None:
