+++ app/routers/guild/webhook/webhook.py
@@ -5,12 +5,9 @@ from fastapi.templating import Jinja2Templates
 
 from typing import List,Dict,Any,Union
 
-from base.aio_req import (
-    aio_get_request,
-    return_permission,
-    discord_oauth_check,
-    get_profile
-)
+from pkg.aio_req import aio_get_request
+from pkg.oauth_check import discord_oauth_check,discord_get_profile
+from pkg.permission import return_permission
 from model_types.discord_type.guild_permission import Permission
 from model_types.discord_type.discord_user_session import DiscordOAuthData
 from model_types.discord_type.discord_type import DiscordUser
@@ -207,7 +204,7 @@ class WebhookView(commands.Cog):
                 # アクセストークンの復号化
                 access_token = session.discord_oauth_data.access_token
                 # Discordのユーザ情報を取得
-                discord_user = await get_profile(access_token=access_token)
+                discord_user = await discord_get_profile(access_token=access_token)
 
                 # トークンが無効
                 if discord_user == None:
