+++ app/routers/guilds.py
@@ -5,10 +5,9 @@ from fastapi.templating import Jinja2Templates
 
 from typing import List,Dict
 
-from base.aio_req import (
-    aio_get_request,
-    discord_oauth_check
-)
+from pkg.aio_req import aio_get_request
+from pkg.oauth_check import discord_oauth_check
+
 from model_types.discord_type.discord_user_session import DiscordOAuthData,MatchGuild
 from model_types.discord_type.discord_type import DiscordUser
